Separate image creds fetching from kubernetes.Cluster
A minor refactor to pull these apart a little and make testing (and
further separation) easier.
